# dashboard/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Memory Observer Dashboard API
  description: |
    REST API for the Memory Observer Dashboard, providing access to workspaces, plans, agents, and events
    from the Project Memory MCP Server.
    
    ## Overview
    
    The Memory Observer Dashboard API allows you to:
    - View and manage MCP workspaces
    - Browse, create, and manage project plans
    - View and deploy agent templates
    - Subscribe to real-time MCP events
    - Search across all workspaces and plans
    
    ## Authentication
    
    Currently, the API runs locally and does not require authentication.
    Future versions may support authentication for remote access.
    
  version: 0.1.0
  contact:
    name: Project Memory MCP
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Local Development Server
...
paths:
...
  /api/agents:
    get:
      tags:
        - Agents
      summary: List all agent templates
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  total:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Agents
      summary: Create a new agent template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request parameters
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: An unexpected error occurred
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
          example: coordinator
        name:
          type: string
          example: Coordinator
        description:
          type: string
        lastModified:
          type: string
          format: date-time
        deployedCount:
          type: integer
        isTemplate:
          type: boolean
...
    CreateAgentRequest:
      type: object
      required:
        - name
        - content
      properties:
        name:
          type: string
          description: Agent name (will be converted to filename)
        content:
          type: string
          description: Markdown content for the agent template
...